+++ Runtime/SDK/Android/Domain/AndroidApphudSubscription.cs
@@ -16,12 +16,13 @@ namespace Apphud.Unity.Android.Domain
             ProductId = _javaObject.Get<string>("productId");
             ExpiresAt = _javaObject.Get<long>("expiresAt");
             StartedAt = _javaObject.Get<long>("startedAt");
-            CanceledAt = _javaObject.GetNullableLong("canceledAt");
+            CanceledAt = _javaObject.GetNullableLong("cancelledAt");
             IsInRetryBilling = _javaObject.Get<bool>("isInRetryBilling");
             IsAutoRenewEnabled = _javaObject.Get<bool>("isAutoRenewEnabled");
             IsIntroductoryActivated = _javaObject.Get<bool>("isIntroductoryActivated");
             Kind = _javaObject.GetEnum<ApphudKind>("kind");
             GroupId = _javaObject.Get<string>("groupId");
+            PurchaseToken = javaObject.Get<string>("purchaseToken");
             IsActive = _javaObject.Call<bool>("isActive");
         }
     }