+++ Runtime/SDK/IOS/Domain/IOSApphudPaywall.cs
@@ -2,7 +2,6 @@
 
 using System.Collections.Generic;
 using Apphud.Unity.Domain;
-using Newtonsoft.Json;
 using UnityEngine.Scripting;
 
 namespace Apphud.Unity.IOS.Domain
@@ -10,15 +9,12 @@ namespace Apphud.Unity.IOS.Domain
     [Preserve]
     internal sealed class IOSApphudPaywallJson
     {
-        public string name;
         public string identifier;
         public string variationName;
         public string experimentName;
-        public string fromPaywall;
-        public List<IOSApphudProductJson> items;
-        public string json;
-
-        [JsonProperty("default")]
+        public string parentPaywallIdentifier;
+        public IOSApphudProductJson[] products;
+        public Dictionary<string, object> json;
         public bool isDefault;
     }
 
@@ -26,22 +22,21 @@ namespace Apphud.Unity.IOS.Domain
     {
         internal IOSApphudPaywall(IOSApphudPaywallJson json, string placementIdentifier)
         {
-            Name = json.name;
             Identifier = json.identifier;
             Default = json.isDefault;
             ExperimentName = json.experimentName;
             VariationName = json.variationName;
-            ParentPaywallIdentifier = json.fromPaywall;
+            ParentPaywallIdentifier = json.parentPaywallIdentifier;
             PlacementIdentifier = placementIdentifier;
 
             Products = new List<ApphudProduct>();
 
-            for (int i = 0; i < json.items.Count; i++)
+            for (int i = 0; i < json.products.Length; i++)
             {
-                Products.Add(new IOSApphudProduct(json.items[i], Identifier, PlacementIdentifier));
+                Products.Add(new IOSApphudProduct(json.products[i], Identifier, PlacementIdentifier));
             }
 
-            Json = json.json != null ? JsonConvert.DeserializeObject<Dictionary<string, object>>(json.json) : null;
+            Json = json.json;
         }
     }
 }