+++ Runtime/SDK/IOS/Domain/IOSApphudPlacement.cs
@@ -1,6 +1,5 @@
 #if UNITY_IOS
 
-using System.Collections.Generic;
 using Apphud.Unity.Domain;
 using UnityEngine.Scripting;
 
@@ -9,7 +8,7 @@ namespace Apphud.Unity.IOS.Domain
     [Preserve]
     internal sealed class IOSApphudPlacementJson
     {
-        public List<IOSApphudPaywallJson> paywalls;
+        public IOSApphudPaywallJson paywall;
         public string identifier;
     }
 
@@ -18,7 +17,7 @@ namespace Apphud.Unity.IOS.Domain
         internal IOSApphudPlacement(IOSApphudPlacementJson json)
         {
             Identifier = json.identifier;
-            Paywall = json.paywalls.Count > 0 ? new IOSApphudPaywall(json.paywalls[0], Identifier) : null;
+            Paywall = json.paywall != null ? new IOSApphudPaywall(json.paywall, Identifier) : null;
         }
     }
 }