+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.2.0] - 2024-10-15
+
+- New methods
+  - ForceFlushUserProperties
+  - DeferPlacements
+- Modified methods
+  - RefreshUserData
+- Fixed methods
+  - GrantPromotional
+- Added SKProduct to ApphudProduct on iOS
+- Added PurchaseToken to ApphudSubscription and ApphudNonRenewingPurchase
+
 ## [1.1.0] - 2024-07-29
 
 - Observer mode